feat(app): implement backend fallbacks for backends that do not support the image format
